+++ lua/github-theme/group/editor.lua
@@ -1,3 +1,5 @@
+local C = require('github-theme.lib.color')
+
 local M = {}
 
 function M.get(spec, config)
@@ -17,6 +19,7 @@ function M.get(spec, config)
   end
 
   local c = spec.palette
+  local sts_bg = C.from_hex(spec.bg0):blend(C.from_hex(c.blue.base), 0.7):to_css()
 
  -- stylua: ignore start
  local groups = {
@@ -63,7 +66,7 @@ function M.get(spec, config)
     Pmenu           = { fg = spec.fg1, bg = spec.bg0 }, -- Popup menu: normal item.
     PmenuSel        = { bg = spec.sel1 }, -- Popup menu: selected item.
     PmenuSbar       = { link = 'Pmenu' }, -- Popup menu: scrollbar.
-    PmenuThumb      = { bg = spec.sel1 }, -- Popup menu: Thumb of the scrollbar.
+    PmenuThumb      = { bg = spec.sel0 }, -- Popup menu: Thumb of the scrollbar.
     Question        = { link = 'MoreMsg' }, -- |hit-enter| prompt and yes/no questions
     QuickFixLine    = { link = 'CursorLine' }, -- Current |quickfix| item in the quickfix window. Combined with |hl-CursorLine| when the cursor is there.
 
@@ -76,7 +79,7 @@ function M.get(spec, config)
     SpellCap        = { sp = spec.diag.warn, style = 'undercurl' }, -- Word that should start with a capital. |spell| Combined with the highlighting used otherwise.
     SpellLocal      = { sp = spec.diag.info, style = 'undercurl' }, -- Word that is recognized by the spellchecker as one that is used in another region. |spell| Combined with the highlighting used otherwise.
     SpellRare       = { sp = spec.diag.info, style = 'undercurl' }, -- Word that is recognized by the spellchecker as one that is hardly ever used.  |spell| Combined with the highlighting used otherwise.
-    StatusLine      = { fg = spec.fg2, bg = spec.bg0 }, -- status line of current window
+    StatusLine      = { fg = spec.bg0, bg = sts_bg }, -- status line of current window
     StatusLineNC    = { fg = spec.fg3, bg = spec.bg0 }, -- status lines of not-current windows Note: if this is equal to 'StatusLine' Vim will use '^^^' in the status line of the current window.
 
     TabLine         = { fg = spec.fg2, bg = spec.bg2 }, -- tab pages line, not active tab page label